+++ integration-tests/src/test/java/org/ihtsdo/otf/mapping/test/other/WorkflowPathStatesNonLegacyPathTest.java
@@ -91,8 +91,11 @@ public class WorkflowPathStatesNonLegacyPathTest {
       mappingService.removeMapProject(mp.getId());
 
     for (MapUser mu : mappingService.getMapUsers().getIterable())
-      mappingService.removeMapUser(mu.getId());
-
+      if (!mu.getUserName().equals("guest")
+          && !mu.getUserName().equals("loader")
+          && !mu.getUserName().equals("qa")) {
+        mappingService.removeMapUser(mu.getId());
+      }
     for (TrackingRecord tr : workflowService.getTrackingRecords().getIterable())
       workflowService.removeTrackingRecord(tr.getId());
